책 정보 조회 및 결과 처리 방식 변경

+++ data/src/main/java/com/mj/data/BookRepositoryImpl.kt
@@ -1,11 +1,10 @@
 package com.mj.data
 
-import com.mj.data.mapper.mapperToBook
 import com.mj.data.datasource.BookRemoteDataSource
+import com.mj.data.mapper.mapperToBook
 import com.mj.domain.model.Book
 import com.mj.domain.repository.BookRepository
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
 
@@ -13,7 +12,7 @@ class BookRepositoryImpl @Inject constructor(
     private val remoteDateSource: BookRemoteDataSource,
 ) : BookRepository {
 
-    override suspend fun searchBook(query: String, currentIndex: Long): Flow<Book> = flow {
+    override suspend fun searchBook(query: String, currentIndex: Int): Flow<Book> = flow {
         remoteDateSource.requestBooksData(query, currentIndex).collect {
             emit(it.mapperToBook())
         }
