+++ data/src/main/java/com/mj/data/datasource/BookRemoteDataSource.kt
@@ -4,5 +4,5 @@ import com.mj.data.model.BookResponse
 import kotlinx.coroutines.flow.Flow
 
 interface BookRemoteDataSource {
-    suspend fun requestBooksData(query: String, currentIndex: Long): Flow<BookResponse>
+    suspend fun requestBooksData(query: String, currentIndex: Int): Flow<BookResponse>
 }
