+++ domain/src/main/java/com/mj/domain/repository/BookRepository.kt
@@ -4,5 +4,5 @@ import com.mj.domain.model.Book
 import kotlinx.coroutines.flow.Flow
 
 interface BookRepository {
-    suspend fun searchBook(query: String, currentIndex: Long): Flow<Book>
+    suspend fun searchBook(query: String, currentIndex: Int): Flow<Book>
 }
